Created class User, function checkPermission and added aa system of user groups
The user class's prototype is located in /controllers/User.php it stores the information of the user that is in the file, and calculates its forbidden files with its groups.

the function checkPermission uses this object to check if the targeted URL makes part of the forbidden list of files.

I also made a forbidden access page and changes the HTTP response codes